Improved AI Education Considerably.
AI actually picks good guardians for their children based on the education focus of the child and the traits and stats of the guardian.

AI children that have a martial education focus but whos gender is not permitted to actually become knights or commanders will have their focus swapped to one more appropriate to their traits.